Moved getPluginConfig to utils
